+++ cmsplugin_cascade/clipboard/admin.py
@@ -8,14 +8,17 @@ from django.forms.utils import flatatt
 from django.utils.encoding import force_text
 from django.utils.html import format_html
 from django.utils.translation import ugettext_lazy as _
+
 from cms.api import add_plugin
-from cms.models.placeholdermodel import Placeholder
 from cms.models.placeholderpluginmodel import PlaceholderReference
 from cms.plugin_pool import plugin_pool
 from cms.utils import get_language_from_request
+
 from jsonfield.fields import JSONField
 from djangocms_text_ckeditor.models import Text
-from cmsplugin_cascade.models import CascadeClipboard
+from djangocms_text_ckeditor.utils import plugin_tags_to_id_list, replace_plugin_tags
+
+from cmsplugin_cascade.models import CascadeElement, CascadeClipboard
 
 
 class JSONAdminWidget(widgets.Textarea):
@@ -72,7 +75,7 @@ class CascadeClipboardAdmin(admin.ModelAdmin):
             request.POST['_continue'] = True
         super(CascadeClipboardAdmin, self).save_model(request, obj, form, change)
         if request.POST.get('restore_clipboard'):
-            self._deserialize_clipboard(language, obj.data)
+            self._deserialize_clipboard(request, obj.data)
 
     def _serialize_clipboard(self, language):
         """
@@ -80,12 +83,13 @@ class CascadeClipboardAdmin(admin.ModelAdmin):
         """
         def populate_data(parent, data):
             for child in plugin_qs.filter(parent=parent).order_by('position'):
-                instance, dummy = child.get_plugin_instance(self.admin_site)
+                instance, plugin = child.get_plugin_instance(self.admin_site)
+                plugin_type = plugin.__class__.__name__
                 try:
-                    entry = (child.plugin_type, instance.get_data_representation(), [])
+                    entry = (plugin_type, plugin.get_data_representation(instance), [])
                 except AttributeError:
                     if isinstance(instance, Text):
-                        entry = (child.plugin_type, {'body': instance.body}, [])
+                        entry = (plugin_type, {'body': instance.body}, [])
                     else:
                         continue
                 data.append(entry)
@@ -99,19 +103,45 @@ class CascadeClipboardAdmin(admin.ModelAdmin):
             populate_data(None, data['plugins'])
         return data
 
-    def _deserialize_clipboard(self, language, data):
+    def _deserialize_clipboard(self, request, data):
         """
         Restore clipboard by creating plugins from given data.
         """
-        def plugins_from_data(parent, data):
+        def plugins_from_data(placeholder, parent, data):
             for entry in data:
                 plugin_type = plugin_pool.get_plugin(entry[0])
                 kwargs = dict(entry[1])
-                instance = add_plugin(clipboard, plugin_type, language, target=parent, **kwargs)
+                inlines = kwargs.pop('inlines', [])
+                shared_glossary = kwargs.pop('shared_glossary', None)
+                instance = add_plugin(placeholder, plugin_type, language, target=parent, **kwargs)
+                if isinstance(instance, CascadeElement):
+                    instance.plugin_class.add_inline_elements(instance, inlines)
+                    instance.plugin_class.add_shared_reference(instance, shared_glossary)
+
                 # for some unknown reasons add_plugin sets instance.numchild 0,
                 # but fixing and save()-ing 'instance' executes some filters in an unwanted manner
-                plugins_from_data(instance, entry[2])
+                plugins_from_data(placeholder, instance, entry[2])
 
-        clipboard = Placeholder.objects.filter(slot='clipboard').last()
-        clipboard.cmsplugin_set.all().delete()
-        plugins_from_data(None, data['plugins'])
+                if isinstance(instance, Text):
+                    # we must convert the old plugin IDs into the new ones,
+                    # otherwise links are not displayed
+                    id_dict = dict(zip(
+                        plugin_tags_to_id_list(instance.body),
+                        (t[0] for t in instance.get_children().values_list('id'))
+                    ))
+                    instance.body = replace_plugin_tags(instance.body, id_dict)
+                    instance.save()
+
+        language = get_language_from_request(request)
+
+        clipboard = request.toolbar.clipboard
+        ref_plugin = clipboard.cmsplugin_set.first()
+        if ref_plugin is None:
+            # the clipboard is empty
+            root_plugin = add_plugin(clipboard, 'PlaceholderPlugin', language, name='clipboard')
+        else:
+            # remove old entries from the clipboard
+            root_plugin = ref_plugin.cms_placeholderreference
+            inst = ref_plugin.get_plugin_instance()[0]
+            inst.placeholder_ref.get_plugins().delete()
+        plugins_from_data(root_plugin.placeholder_ref, None, data['plugins'])