+++ cmsplugin_cascade/plugin_base.py
@@ -2,27 +2,35 @@
 from __future__ import unicode_literals
 
 from collections import OrderedDict
+from distutils.version import LooseVersion
+
 from django.core.exceptions import ImproperlyConfigured, ObjectDoesNotExist
-from django.forms.widgets import media_property
+from django.forms import MediaDefiningClass
 from django.utils import six
 from django.utils.functional import lazy
 from django.utils.module_loading import import_string
 from django.utils.translation import string_concat
 from django.utils.safestring import SafeText, mark_safe
-from cms.plugin_pool import plugin_pool
+
+from cms import __version__ as cms_version
 from cms.plugin_base import CMSPluginBaseMetaclass, CMSPluginBase
-from cms.utils.placeholder import get_placeholder_conf
 from cms.utils.compat.dj import is_installed
+
 from . import settings
 from .fields import GlossaryField
-from .mixins import TransparentMixin
 from .models_base import CascadeModelBase
 from .models import CascadeElement, SharableCascadeElement
 from .generic.mixins import SectionMixin, SectionModelMixin
 from .sharable.forms import SharableGlossaryMixin
 from .extra_fields.mixins import ExtraFieldsMixin
 from .widgets import JSONMultiWidget
+from .hide_plugins import HidePluginMixin
 from .render_template import RenderTemplateMixin
+from .utils import remove_duplicates
+
+mark_safe_lazy = lazy(mark_safe, six.text_type)
+
+fake_proxy_models = {}
 
 
 def create_proxy_model(name, model_mixins, base_model, attrs=None, module=None):
@@ -31,27 +39,31 @@ def create_proxy_model(name, model_mixins, base_model, attrs=None, module=None):
     """
     class Meta:
         proxy = True
-        # using a dummy name prevents `makemigrations` to create a model migration
-        app_label = 'cascade_dummy'
+        app_label = 'cmsplugin_cascade'
 
     name = str(name + 'Model')
     bases = model_mixins + (base_model,)
-    attrs = {} if attrs is None else dict(attrs)
-    try:
-        attrs.update(Meta=Meta, __module__=module)
-        Model = type(name, bases, attrs)
-    except RuntimeError:
-        Meta.app_label = 'cascade_dummy_dummy'
-        attrs.update(Meta=Meta, __module__=module)
-        Model = type(name, bases, attrs)
+    attrs = dict(attrs or {}, Meta=Meta, __module__=module)
+    Model = type(name, bases, attrs)
+    fake_proxy_models[name] = bases
     return Model
 
-mark_safe_lazy = lazy(mark_safe, six.text_type)
 
+class CascadePluginMixinMetaclass(MediaDefiningClass):
+    ring_plugin_bases = {}
 
-class CascadePluginMixinMetaclass(type):
     def __new__(cls, name, bases, attrs):
         cls.build_glossary_fields(bases, attrs)
+        ring_plugin = attrs.get('ring_plugin')
+        if ring_plugin:
+            ring_plugin_bases = [b.ring_plugin for b in bases
+                                 if hasattr(b, 'ring_plugin') and b.ring_plugin != ring_plugin]
+
+            # remember the dependencies
+            cls.ring_plugin_bases.setdefault(ring_plugin, [])
+            cls.ring_plugin_bases[ring_plugin].extend(ring_plugin_bases)
+            cls.ring_plugin_bases[ring_plugin] = remove_duplicates(cls.ring_plugin_bases[ring_plugin])
+
         new_class = super(CascadePluginMixinMetaclass, cls).__new__(cls, name, bases, attrs)
         return new_class
 
@@ -110,26 +122,30 @@ class CascadePluginBaseMetaclass(CascadePluginMixinMetaclass, CMSPluginBaseMetac
     plugins_with_bookmark = list(settings.CMSPLUGIN_CASCADE['plugins_with_bookmark'])
     plugins_with_sharables = dict(settings.CMSPLUGIN_CASCADE['plugins_with_sharables'])
     plugins_with_extra_render_templates = settings.CMSPLUGIN_CASCADE['plugins_with_extra_render_templates'].keys()
+    allow_plugin_hiding = settings.CMSPLUGIN_CASCADE['allow_plugin_hiding']
+    exclude_hiding_plugin = list(settings.CMSPLUGIN_CASCADE['exclude_hiding_plugin'])
 
     def __new__(cls, name, bases, attrs):
         model_mixins = attrs.pop('model_mixins', ())
+        if (cls.allow_plugin_hiding and name not in cls.exclude_hiding_plugin and 'name' in attrs and
+            not attrs.get('text_enabled')):
+            bases = (HidePluginMixin,) + bases
         if name in cls.plugins_with_extra_fields:
-            ExtraFieldsMixin.media = media_property(ExtraFieldsMixin)
             bases = (ExtraFieldsMixin,) + bases
         if name in cls.plugins_with_bookmark:
             bases = (SectionMixin,) + bases
             model_mixins = (SectionModelMixin,) + model_mixins
         if name in cls.plugins_with_sharables:
-            SharableGlossaryMixin.media = media_property(SharableGlossaryMixin)
             bases = (SharableGlossaryMixin,) + bases
             attrs['fields'] = list(attrs.get('fields', ['glossary']))
             attrs['fields'].extend([('save_shared_glossary', 'save_as_identifier'), 'shared_glossary'])
             attrs['sharable_fields'] = cls.plugins_with_sharables[name]
             base_model = SharableCascadeElement
         else:
+            attrs['exclude'] = list(attrs.get('exclude', []))
+            attrs['exclude'].append('shared_glossary')
             base_model = CascadeElement
         if name in cls.plugins_with_extra_render_templates:
-            RenderTemplateMixin.media = media_property(RenderTemplateMixin)
             bases = (RenderTemplateMixin,) + bases
         if name == 'SegmentPlugin':
             # SegmentPlugin shall additionally inherit from configured mixin classes
@@ -144,7 +160,69 @@ class CascadePluginBaseMetaclass(CascadePluginMixinMetaclass, CMSPluginBaseMetac
         if 'name' in attrs and settings.CMSPLUGIN_CASCADE['plugin_prefix']:
             attrs['name'] = mark_safe_lazy(string_concat(
                 settings.CMSPLUGIN_CASCADE['plugin_prefix'], "&nbsp;", attrs['name']))
+
         return super(CascadePluginBaseMetaclass, cls).__new__(cls, name, bases, attrs)
+
+
+class TransparentWrapper(object):
+    """
+    Add this mixin class to other Cascade plugins, wishing to be added transparently between other
+    plugins restricting parent-children relationships.
+    For instance: A BootstrapColumnPlugin can only be added as a child to a RowPlugin. This means
+    that no other wrapper can be added between those two plugins. By adding this mixin class we can
+    allow any plugin to behave transparently, just as if it would not have be inserted into the DOM
+    tree. When moving plugins in- and out of transparent wrapper plugins, always reload the page, so
+    that the parent-children relationships can be updated.
+    """
+    child_plugins_cache = False
+    parent_plugins_cache = False
+
+    @classmethod
+    def get_child_classes(cls, slot, page, instance=None):
+        if hasattr(cls, 'direct_child_classes'):
+            return cls.direct_child_classes
+        while True:
+            instance = instance.get_parent_instance()
+            if instance is None:
+                return super(TransparentWrapper, cls).get_child_classes(slot, page, instance)
+            if not issubclass(instance.plugin_class, TransparentWrapper):
+                return instance.plugin_class.get_child_classes(slot, page, instance)
+
+    @classmethod
+    def get_parent_classes(cls, slot, page, instance=None):
+        if hasattr(cls, 'direct_parent_classes'):
+            return cls.direct_parent_classes
+        parent_classes = set(super(TransparentWrapper, cls).get_parent_classes(slot, page, instance) or [])
+        if isinstance(instance, CascadeElement):
+            instance = instance.get_parent_instance()
+            if instance is not None:
+                parent_classes.add(instance.plugin_type)
+        return list(parent_classes)
+
+
+class TransparentContainer(TransparentWrapper):
+    """
+    This mixin class marks each plugin inheriting from it, as a transparent container.
+    Such a plugin is added to the global list of entitled parent plugins, which is required if we
+    want to place and move all other Cascade plugins below this container.
+
+    Often, transparent wrapping classes come in pairs. For instance the `AccordionPlugin` containing
+    one or more `PanelPlugin`. Here the `AccordionPlugin` must inherit from `TransparentWrapper`,
+    whereas the `AccordionPlugin` must inherit from the `TransparentContainer`.
+    """
+    @staticmethod
+    def get_plugins():
+        from cms.plugin_pool import plugin_pool
+        global _leaf_transparent_plugins
+
+        try:
+            return _leaf_transparent_plugins
+        except NameError:
+            _leaf_transparent_plugins = [
+                plugin.__name__ for plugin in plugin_pool.get_all_plugins()
+                    if issubclass(plugin, TransparentContainer)
+            ]
+            return _leaf_transparent_plugins
 
 
 class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPluginBase)):
@@ -155,7 +233,8 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
     alien_child_classes = False
 
     class Media:
-        css = {'all': ('cascade/css/admin/partialfields.css', 'cascade/css/admin/editplugin.css',)}
+        css = {'all': ['cascade/css/admin/partialfields.css', 'cascade/css/admin/editplugin.css']}
+        js = ['cascade/js/underscore.js', 'cascade/js/ring.js']
 
     def __init__(self, model=None, admin_site=None, glossary_fields=None):
         super(CascadePluginBase, self).__init__(model, admin_site)
@@ -164,31 +243,45 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         elif not hasattr(self, 'glossary_fields'):
             self.glossary_fields = []
 
-    def get_parent_classes(self, slot, page):
-        template = page and page.get_template() or None
-        ph_conf = get_placeholder_conf('parent_classes', slot, template, default={})
-        parent_classes = ph_conf.get(self.__class__.__name__, self.parent_classes)
+    @classmethod
+    def _get_parent_classes_transparent(cls, slot, page, instance=None):
+        """
+        Return all parent classes including those marked as "transparent".
+        """
+        parent_classes = super(CascadePluginBase, cls).get_parent_classes(slot, page, instance)
         if parent_classes is None:
-            return
-        # allow all parent classes which inherit from TransparentMixin
-        parent_classes = set(parent_classes)
-        for p in plugin_pool.get_all_plugins():
-            if self.allow_children and issubclass(p, TransparentMixin):
-                parent_classes.add(p.__name__)
-        return tuple(parent_classes)
+            if cls.get_require_parent(slot, page) is False:
+                return
+            parent_classes = []
 
-    def get_child_classes(self, slot, page):
-        if isinstance(self.child_classes, (list, tuple)):
-            return self.child_classes
-        # otherwise determine child_classes by evaluating parent_classes from other plugins
+        # add all plugins marked as 'transparent', since they all are potential parents
+        parent_classes = set(parent_classes)
+        parent_classes.update(TransparentContainer.get_plugins())
+        return list(parent_classes)
+
+    @classmethod
+    def get_child_classes(cls, slot, page, instance=None):
+        plugin_type = cls.__name__
         child_classes = set()
-        for p in plugin_pool.get_all_plugins():
-            if (isinstance(p.parent_classes, (list, tuple)) and self.__class__.__name__ in p.parent_classes or
-              p.parent_classes is None and issubclass(p, CascadePluginBase) or
-              isinstance(self.alien_child_classes, (list, tuple)) and p.__name__ in self.alien_child_classes or
-              self.alien_child_classes is True and p.__name__ in settings.CMSPLUGIN_CASCADE['alien_plugins']):
-                child_classes.add(p.__name__)
-        return tuple(child_classes)
+        for child_class in cls.get_child_plugin_candidates(slot, page):
+            if issubclass(child_class, CascadePluginBase):
+                own_child_classes = getattr(cls, 'child_classes', None) or []
+                child_parent_classes = child_class._get_parent_classes_transparent(slot, page, instance)
+                if isinstance(child_parent_classes, (list, tuple)) and plugin_type in child_parent_classes:
+                    child_classes.add(child_class)
+                elif plugin_type in own_child_classes:
+                    child_classes.add(child_class)
+                elif child_parent_classes is None:
+                    child_classes.add(child_class)
+            else:
+                if cls.alien_child_classes and child_class.__name__ in settings.CMSPLUGIN_CASCADE['alien_plugins']:
+                    child_classes.add(child_class)
+
+        return list(cc.__name__ for cc in child_classes)
+
+    @classmethod
+    def get_parent_classes(cls, slot, page, instance=None):
+        return cls._get_parent_classes_transparent(slot, page, instance)
 
     @classmethod
     def get_identifier(cls, instance):
@@ -255,6 +348,25 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
             instance.glossary = {}
         return False
 
+    @classmethod
+    def get_data_representation(cls, instance):
+        """
+        Return a representation of the given instance suitable for a serialized representation.
+        """
+        return {'glossary': instance.glossary}
+
+    @classmethod
+    def add_inline_elements(cls, instance, inlines):
+        """
+        Hook to create (sortable) inline elements for the given instance.
+        """
+
+    @classmethod
+    def add_shared_reference(cls, instance, shared_glossary):
+        """
+        Hook to add a reference pointing onto an existing SharedGlossary instance.
+        """
+
     def extend_children(self, parent, wanted_children, child_class, child_glossary=None):
         """
         Extend the number of children so that the parent object contains wanted children.
@@ -281,8 +393,11 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         form = super(CascadePluginBase, self).get_form(request, obj, **kwargs)
         # help_text can not be cleared using an empty string in modelform_factory
         form.base_fields['glossary'].help_text = ''
-        for field in glossary_fields:
-            form.base_fields['glossary'].validators.append(field.run_validators)
+        if request.method == 'POST':
+            is_shared = bool(request.POST.get('shared_glossary'))
+            for field in glossary_fields:
+                if not (is_shared and field.name in self.sharable_fields):
+                    form.base_fields['glossary'].validators.append(field.run_validators)
         form.glossary_fields = glossary_fields
         return form
 
@@ -353,12 +468,18 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
         return None, None
 
     def render_change_form(self, request, context, add=False, change=False, form_url='', obj=None):
+        ring_plugin_bases = dict((ring_plugin, ['django.cascade.{}'.format(b) for b in bases])
+                                 for ring_plugin, bases in CascadePluginMixinMetaclass.ring_plugin_bases.items())
         context.update(
-            base_plugins=['django.cascade.{}'.format(b) for b in self.get_ring_bases()],
+            ring_plugin_bases=ring_plugin_bases,
             plugin_title=string_concat(self.module, " ", self.name, " Plugin"),
             plugin_intro=mark_safe(getattr(self, 'intro_html', '')),
             plugin_footnote=mark_safe(getattr(self, 'footnote_html', '')),
         )
+        if hasattr(self, 'ring_plugin'):
+            context.update(
+                ring_plugin=self.ring_plugin,
+            )
 
         # remove glossary field from rendered form
         form = context['adminform'].form
@@ -370,9 +491,15 @@ class CascadePluginBase(six.with_metaclass(CascadePluginBaseMetaclass, CMSPlugin
             pass
         return super(CascadePluginBase, self).render_change_form(request, context, add, change, form_url, obj)
 
-    def get_ring_bases(self):
+    def in_edit_mode(self, request, placeholder):
         """
-        Hook to return a list of base plugins required to build the JavaScript counterpart for the
-        current plugin. The named JavaScript plugin must have been created using ``ring.create``.
+        Returns True, if the plugin is in "edit mode".
         """
-        return []
+        toolbar = getattr(request, 'toolbar', None)
+        edit_mode = getattr(toolbar, 'edit_mode', False) and getattr(placeholder, 'is_editable', True)
+        if edit_mode:
+            if LooseVersion(cms_version) < LooseVersion('3.4.0'):
+                edit_mode = placeholder.has_change_permission(request)
+            else:
+                edit_mode = placeholder.has_change_permission(request.user)
+        return edit_mode