+++ cmsplugin_cascade/settings.py
@@ -6,11 +6,18 @@ from collections import OrderedDict
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.utils.translation import ugettext_lazy as _
+
 from cmsplugin_cascade.extra_fields.config import PluginExtraFieldsConfig
-from cmsplugin_cascade.widgets import MultipleCascadingSizeWidget, ColorPickerWidget, SelectOverflowWidget
+from cmsplugin_cascade.widgets import (
+    NumberInputWidget, CascadingSizeWidget, MultipleCascadingSizeWidget, ColorPickerWidget,
+    SelectOverflowWidget)
 
 
-CASCADE_PLUGINS = getattr(settings, 'CMSPLUGIN_CASCADE_PLUGINS', ('cmsplugin_cascade.generic', 'cmsplugin_cascade.link',))
+CASCADE_PLUGINS = getattr(settings, 'CMSPLUGIN_CASCADE_PLUGINS', (
+    'cmsplugin_cascade.generic',
+    'cmsplugin_cascade.icon',
+    'cmsplugin_cascade.link',
+))
 
 CMSPLUGIN_CASCADE = getattr(settings, 'CMSPLUGIN_CASCADE', {})
 orig_config = dict(CMSPLUGIN_CASCADE)
@@ -19,42 +26,36 @@ orig_config = dict(CMSPLUGIN_CASCADE)
 if not isinstance(CMSPLUGIN_CASCADE.get('plugins_with_extra_fields', {}), dict):
     raise ImproperlyConfigured("CMSPLUGIN_CASCADE['plugins_with_extra_fields'] must be declared as dict.")
 
-CMSPLUGIN_CASCADE.setdefault('fontawesome_css_url', 'https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css')
-"""Specify location of CSS file used to load fonts from http://fontawesome.io/"""
-
 CMSPLUGIN_CASCADE.setdefault('alien_plugins', ['TextPlugin'])
 
 # Use a prefix to symbolize a Cascade plugin in case there are ambiguous names.
 CMSPLUGIN_CASCADE.setdefault('plugin_prefix', None)
 
-CMSPLUGIN_CASCADE['dependencies'] = {
-    'cascade/js/ring.js': 'cascade/js/underscore.js',
-    'cascade/js/admin/sharableglossary.js': 'cascade/js/ring.js',
-    'cascade/js/admin/segmentplugin.js': 'cascade/js/ring.js',
-    'cascade/js/admin/jumbotronplugin.js': 'cascade/js/ring.js',
-    'cascade/js/admin/fonticonplugin.js': 'cascade/js/ring.js',
-    'cascade/js/admin/linkpluginbase.js': ('cascade/js/admin/sharableglossary.js',),
-    'cascade/js/admin/defaultlinkplugin.js': ('cascade/js/admin/linkpluginbase.js',),
-    'cascade/js/admin/imageplugin.js': ('cascade/js/admin/linkpluginbase.js',),
-    'cascade/js/admin/pictureplugin.js': ('cascade/js/admin/linkpluginbase.js',),
-}
-"""The editor of some plugins requires JavaScript file. Here we can specify which is a list of dependencies"""
-CMSPLUGIN_CASCADE['dependencies'].update(orig_config.get('dependencies', {}))
-
 if 'cmsplugin_cascade.extra_fields' in settings.INSTALLED_APPS:
     CMSPLUGIN_CASCADE['plugins_with_extra_fields'] = {
         'BootstrapButtonPlugin': PluginExtraFieldsConfig(),
         'BootstrapRowPlugin': PluginExtraFieldsConfig(),
-        'BootstrapJumbotronPlugin': PluginExtraFieldsConfig(inline_styles={
-            'extra_fields:Paddings': ['padding-top', 'padding-bottom'],
-            'extra_units:Paddings': 'px,em'}),
+        'BootstrapJumbotronPlugin': PluginExtraFieldsConfig(
+            inline_styles={
+                'extra_fields:Paddings': ['padding-top', 'padding-bottom'],
+                'extra_units:Paddings': 'px,em'
+            }
+        ),
         'SimpleWrapperPlugin': PluginExtraFieldsConfig(),
-        'HeadingPlugin': PluginExtraFieldsConfig(inline_styles={
-            'extra_fields:Paddings': ['margin-top', 'margin-right', 'margin-bottom', 'margin-left'],
-            'extra_units:Paddings': 'px,em'}, allow_override=False),
-        'HorizontalRulePlugin': PluginExtraFieldsConfig(inline_styles={
-            'extra_fields:Paddings': ['margin-top', 'margin-bottom'],
-            'extra_units:Paddings': 'px,em'}, allow_override=False),
+        'HeadingPlugin': PluginExtraFieldsConfig(
+            inline_styles={
+                'extra_fields:Margins': ['margin-top', 'margin-right', 'margin-bottom', 'margin-left'],
+                'extra_units:Margins': 'px,em'
+            },
+            allow_override=False
+        ),
+        'HorizontalRulePlugin': PluginExtraFieldsConfig(
+            inline_styles={
+                'extra_fields:Paddings': ['margin-top', 'margin-bottom'],
+                'extra_units:Paddings': 'px,em'
+            },
+            allow_override=False
+        ),
     }
     CMSPLUGIN_CASCADE['plugins_with_extra_fields'].update(
         orig_config.get('plugins_with_extra_fields', {}))
@@ -70,6 +71,13 @@ styles.
 """
 
 CMSPLUGIN_CASCADE.setdefault('plugins_with_sharables', {})
+CMSPLUGIN_CASCADE['plugins_with_sharables'].setdefault(
+    'FramedIconPlugin', ('font_size', 'color', 'background_color', 'text_align', 'border', 'border_radius'),
+)
+
+CMSPLUGIN_CASCADE.setdefault('exclude_hiding_plugin', (
+    'SegmentPlugin',
+))
 
 CMSPLUGIN_CASCADE.setdefault('link_plugin_classes', (
     'cmsplugin_cascade.link.plugin_base.DefaultLinkPluginBase',
@@ -92,6 +100,7 @@ CMSPLUGIN_CASCADE['extra_inline_styles'] = OrderedDict((
     ('Widths', (('min-width', 'width', 'max-width',), MultipleCascadingSizeWidget)),
     ('Heights', (('min-height', 'height', 'max-height',), MultipleCascadingSizeWidget)),
     ('Font Size', (('font-size',), MultipleCascadingSizeWidget)),
+    ('Line Height', (('line-height',), NumberInputWidget)),
     ('Colors', (('color', 'background-color',), ColorPickerWidget)),
     ('Overflow', (('overflow', 'overflow-x', 'overflow-y',), SelectOverflowWidget)),
 ))
@@ -108,6 +117,16 @@ CMSPLUGIN_CASCADE.setdefault('plugins_with_extra_render_templates', {
     )
 })
 
+CMSPLUGIN_CASCADE.setdefault('allow_plugin_hiding', False)
+"""
+Add a checkbox to each named plugin, which if checked hides that plugin during the rendering phase.
+Useful to temporarily deactivate plugins, instead of removing them.
+"""
+
 # Folder where extracted icon fonts are stored.
 CMSPLUGIN_CASCADE.setdefault('icon_font_root',
                              os.path.abspath(os.path.join(settings.MEDIA_ROOT, 'icon_fonts')))
+
+CSS_PREFIXES = {
+    'image_set': ['-webkit-image-set', '-moz-image-set', '-o-image-set', '-ms-image-set', 'image-set'],
+}