+++ docs/source/changelog.rst
@@ -4,6 +4,49 @@
 Release History
 ===============
 
+0.12.0
+------
+* Added compatibility for Django version 1.10.
+* Added compatibility for django-CMS version 3.4.
+* Added monkey patch to resolve issues handled by PR https://github.com/divio/django-cms/pull/5809
+* Added compatibility for djangocms-text-ckeditor-3.4.
+* **Important for AngularJS users**: Please upgrade to angular-ui-bootstrap version 0.14.3. All
+  versions later than 0.13 use the prefix ``uib-`` on all AngularJS directives, hence this upgrade
+  is required.
+* In the ``CarouselSlide`` plugin, caption is added as a child ``TextPlugin`` instead of using the
+  glossary. Currently the migration of ``TextLinkPlugins`` inside this caption field does not work
+  properly. Please create an issue, if you really need it.
+* Added method ``value_omitted_from_data`` to ``JSONMultiWidget`` to override the Django method
+  implemented in ``django.forms.widgets.MultiWidget``.
+* In ``cmsplugin_cascade.models.CascadeElement`` the foreign key ``shared_glossary`` now is marked
+  as editable. Instead to plugins without sharable glossary, the attribute
+  ``exclude = ['shared_glossary']`` is added.
+* Instead of handling ring.js plugin inheritance through ``get_ring_bases()``, Cascade plugins
+  just have to add ``ring_plugin = '...'`` to their class declaration.
+* Function ``cmsplugin_cascade.utils.resolve_dependencies`` is deprecated, since Javascript
+  dependencies now are handled via their natural inheritance relation.
+* The configuration option ``settings.CMSPLUGIN_CASCADE['dependencies']`` has been removed.
+* Added method ``save()`` to model ``SharedGlossary``, which filters the glossary to be stored to
+  only those fields marked as sharable.
+* Accessing the CMS page via ``plugin_instance.page`` is deprecated and has been replaced by
+  invocations to ``plugin_instance.placeholder.page``.
+* Removed directory ``static/cascade/css/fonts/glyphicons-halflings``, since they are available
+  through the Bootstrap npm packages.
+* All Javascript files accessing a property ``disabled``, now use the proper jQuery function
+  intended for it.
+* Added interface to upload fonts and use them as framed icons, text icons or button decorators.
+* The permission system now is fine grained. Administrators can give their staff users
+  add/change/delete permissions to each of the many Cascade plugins. When adding new plugins, this
+  does not even require a database migration.
+* Fixed: On saving a **CarouselPlugin**, the glossary of it's children, ie. **CarouselSlidePlugin**,
+  is sanitized.
+* Handle the high resolution of the **PicturePlugin** through ``srcset`` rather than a ``@media``
+  query.
+* Handle the high resolution background of the **JumbotronPlugin** through ``image-set`` rather than
+  a ``@media`` query.
+* Use default configurations from provides Cascade settings rathern than from the Django project.
+
+
 0.11.1
 ------
 * Added preconfigured ``FilePathField`` to prevent the creation of useless migration files.